Seperate network layer from minecraft client
