+++ Sources/SwiftCraft/Client/MinecraftClient.swift
@@ -22,14 +22,7 @@ public protocol MinecraftClientProtocol {
 
 /// Baseclass for the library. It manages everything a client have to do.
 open class MinecraftClient: MinecraftClientProtocol {
-    /// The tcp client used to connect to the server
-    private let tcpClient: TCPClientProtocol
-
-    /// A buffer for incomplete incomming minecraft messages.
-    private let incommingDataBuffer = ByteBuffer()
-
-    /// The library of packets used for deserializing.
-    private let packetLibrary: PacketLibrary
+    public private(set) var networkLayer: MinecraftClientNetworkLayer
 
     /// Service to manage the authorization at mojang's session service.
     public let sessionServerService: SessionServerServiceProtocol
@@ -43,50 +36,33 @@ open class MinecraftClient: MinecraftClientProtocol {
     /// The version number used for the protocol
     public private(set) var protocolVersion: Int = 0
 
-    /// Handler to call on open.
-    private var openHandler: (() -> Void)?
-
-    /// Handler to call on close.
-    private var closeHandler: ((ConnectionState) -> Void)?
-
     /// Creates a new `MinecraftClient`
     ///
     /// - Parameters:
     ///   - tcpClient: The client used to connect to the server.
     ///   - packetLibrary: The packet library to use for packet deserializing.
     ///   - sessionServerService: The session server service used for authenticating.
-    public init(tcpClient: TCPClientProtocol,
-                packetLibrary: PacketLibrary,
+    public init(networkLayer: MinecraftClientNetworkLayer,
                 sessionServerService: SessionServerServiceProtocol) {
-        self.tcpClient = tcpClient
-        self.packetLibrary = packetLibrary
+        self.networkLayer = networkLayer
         self.sessionServerService = sessionServerService
-
-        // handle events
-        tcpClient.events { [weak self] event in
-            self?.handleEvent(event)
-        }
+        self.networkLayer.delegate = self
     }
 
-    /// Connect the tcp client to the server
-    public func connect() -> Bool {
-        return tcpClient.connect()
-    }
-
-    /// Connects the tcp client to the server and starts the login
+    /// Connects to the server and starts the login.
     public func connectAndLogin(protocolVersion: Int? = nil) -> Bool {
         guard let protocolVersion = protocolVersion else {
             var reactorID: UUID!
             reactorID = addReactor(ClosureReactor<StatusResponsePacket> { packet, _ in
                 self.removeReactor(with: reactorID)
-                self.close()
+                self.networkLayer.close()
                 _ = self.connectAndLogin(protocolVersion: packet.response.version.protocolVersion)
             })
             return connectAndRequestStatus()
         }
 
         self.protocolVersion = protocolVersion
-        guard connect() else {
+        guard networkLayer.connect() else {
             return false
         }
         connectionState = .handshaking
@@ -98,7 +74,7 @@ open class MinecraftClient: MinecraftClientProtocol {
 
     public func connectAndRequestStatus() -> Bool {
         self.protocolVersion = MinecraftClient.mostRecentSupportedProtocolVersion
-        guard connect() else {
+        guard networkLayer.connect() else {
             return false
         }
         connectionState = .handshaking
@@ -108,40 +84,19 @@ open class MinecraftClient: MinecraftClientProtocol {
         return true
     }
 
-    /// Closes the tcp client connection
-    public func close() {
-        tcpClient.close()
-    }
-
-    /// Sets a handler to be called on connection open.
+    /// Sends a packet to the server.
     ///
-    /// - Parameter handler: The handler to set
-    public func onOpen(handler: (() -> Void)?) {
-        openHandler = handler
+    /// - Parameter packet: The packet to send.
+    public func sendPacket(_ packet: SerializablePacket) {
+        do {
+            if shouldSendPacket(packet, client: self) {
+                try networkLayer.sendPacket(packet)
+                try didSendPacket(packet, client: self)
+            }
+        } catch {
+            print(" ❌ \(error)")
+        }
     }
-
-    /// Sets a handler to be called on connection close.
-    ///
-    /// - Parameter handler: The handler to set
-    public func onClose(handler: ((ConnectionState) -> Void)?) {
-        closeHandler = handler
-    }
-
-    /// The host of the server.
-    public var host: String {
-        return tcpClient.host
-    }
-
-    /// The port of the server.
-    public var port: Int {
-        return tcpClient.port
-    }
-
-    /// Handels the de- and encryption of packets.
-    private var messageCryptor: ContinuousMessageCryptor?
-
-    /// Handels the de- and compression of packets.
-    private var compressor: MessageCompressor?
 
     /// Starts the encryption procedure and session authentication.
     ///
@@ -150,7 +105,7 @@ open class MinecraftClient: MinecraftClientProtocol {
     ///   - publicKey: The servers public key from the `EncryptionRequestPacket`.
     ///   - verifyToken: Teh verification token from the `EncryptionRequestPacket`.
     /// - Throws: Encryption errors.
-    open func startEncryption(serverID: Data, publicKey: Data, verifyToken: Data) throws {
+    public func startEncryption(serverID: Data, publicKey: Data, verifyToken: Data) throws {
         // Generate shared secret used for aes
         let sharedSecret = try CryptoWrapper.generateRandomData(count: 16)
 
@@ -163,35 +118,12 @@ open class MinecraftClient: MinecraftClientProtocol {
             encryptedVerifyToken: Array(encryptedVerifyToken))
 
         // Authenticate with the session server
-        sessionServerService.joinSession(serverID: serverID, sharedSecret: sharedSecret,
+        sessionServerService.joinSession(serverID: serverID,
+                                         sharedSecret: sharedSecret,
                                          publicKey: publicKey) { [weak self] _ in
             self?.sendPacket(responsePacket)
-            self?.enableEncryption(sharedSecret: sharedSecret)
+            self?.networkLayer.enableEncryption(sharedSecret: sharedSecret)
         }
-    }
-
-    /// Enables aes encryption.
-    ///
-    /// - Parameter sharedSecret: The shared secret used for both key and iv.
-    public func enableEncryption(sharedSecret: Data) {
-        messageCryptor = ContinuousMessageCryptor(sharedSecret: sharedSecret)
-    }
-
-    /// Disables encryption.
-    public func disableEncryption() {
-        messageCryptor = nil
-    }
-
-    /// Enables compression for all following packets with at least the size of the threshold.
-    ///
-    /// - Parameter threshold: The minimum size of a packet to compress.
-    public func enableCompression(threshold: Int) {
-        compressor = MessageCompressor(threshold: threshold)
-    }
-
-    /// Disables compression for all following packets.
-    public func disableCompression() {
-        compressor = nil
     }
 
     /// Adds a new `Reactor`. This reactor will be called aber any other reactor added so far.
@@ -215,101 +147,13 @@ open class MinecraftClient: MinecraftClientProtocol {
     public var dimension: Dimension = .overworld
 }
 
-// MARK: - MinecraftClient+Sending
-extension MinecraftClient {
-    /// Sends raw bytes to the server. This function should not be called manually when sending packets.
-    ///
-    /// - Parameter bytes: The bytes to send.
-    public func send(_ bytes: ByteArray) throws {
-        try tcpClient.send(bytes: encryptMessageIfRequired(bytes))
+extension MinecraftClient: MinecraftClientNetworkLayerDelegate {
+    public var serializationContext: SerializationContext {
+        return self
     }
 
-    /// Sends a packet to the server.
-    ///
-    /// - Parameter packet: The packet to send.
-    public func sendPacket(_ packet: SerializablePacket) {
-        do {
-            if shouldSendPacket(packet, client: self) {
-                guard let serializedPacket = packet.serialize(context: self) else {
-                    print(" ❌ packet cannot be encoded \(packet)")
-                    return
-                }
-                let compressedMessage = try compressMessageIfRequired(serializedPacket)
-                let messageSize = VarInt32(compressedMessage.count).directSerialized()
-
-                try send(messageSize + compressedMessage)
-                didSendPacket(packet, client: self)
-            }
-        } catch {
-            print(" ❌ \(error)")
-        }
-    }
-}
-
-// MARK: - MinecraftClient+Receiving
-extension MinecraftClient {
-    /// Handels a new `TCPClientEvent`.
-    ///
-    /// - Parameter event: The event to handle.
-    private func handleEvent(_ event: TCPClientEvent) {
-        switch event {
-        case let .received(data):
-            handleMessage(Array(data))
-        case .streamOpened:
-            openHandler?()
-        case .streamClosed:
-            closeHandler?(connectionState)
-        }
-    }
-
-    /// Handels a new incomming message.
-    ///
-    /// - Parameter bytes: The message to handle.
-    private func handleMessage(_ bytes: ByteArray) {
-        let decryptedBytes = decryptMessageIfRequired(bytes)
-        incommingDataBuffer.write(elements: decryptedBytes)
-
-        while true {
-            // Read next packet length
-            guard let packetLength = try? VarInt32(from: incommingDataBuffer).integer else {
-                incommingDataBuffer.resetPosition()
-                return
-            }
-
-            // Read next packet data
-            guard let packetData = try? incommingDataBuffer.read(lenght: packetLength) else {
-                incommingDataBuffer.resetPosition()
-                return
-            }
-
-            // Drop read bytes.
-            incommingDataBuffer.dropReadElements()
-
-            // Handle packet
-            do {
-                try handlePacketData(packetData)
-            } catch {
-                if case let PacketLibraryError.unknowPacketID(packetID) = error {
-                    let hexID = String(packetID.id, radix: 16, uppercase: true)
-                    print(" ❗️ unknown packet: \(packetID.connectionState): 0x\(hexID)")
-                } else {
-                    print(" ❌ \(error)")
-                }
-            }
-        }
-    }
-
-    /// Handels the data of a new packet. This includes deserializing and reactors.
-    ///
-    /// - Parameter packetData: The dat of the packet.
-    /// - Throws: Any error which might occure.
-    private func handlePacketData(_ packetData: ByteArray) throws {
-        let packetBuffer = try Buffer(elements: decompressMessageIfRequired(packetData))
-        let rawPacketID = try VarInt32(from: packetBuffer).integer
-        let packetID = connectionState.packetID(with: rawPacketID)
-
-        let packet = try packetLibrary.parse(packetBuffer, packetID: packetID, context: self)
-        try self.didReceivedPacket(packet, client: self)
+    public func didReceivePacket(_ packet: DeserializablePacket) {
+        try? didReceivedPacket(packet, client: self)
     }
 }
 
@@ -327,57 +171,9 @@ extension MinecraftClient: Reactor {
         }
     }
 
-    public func didSendPacket(_ packet: SerializablePacket, client: MinecraftClient) {
-        reactors.forEach {
-            $0.value.didSendPacket(packet, client: client)
+    public func didSendPacket(_ packet: SerializablePacket, client: MinecraftClient) throws {
+        try reactors.forEach {
+            try $0.value.didSendPacket(packet, client: client)
         }
-    }
-}
-
-// MARK: - MinecraftClient+Enryption
-extension MinecraftClient {
-    /// Encrypts a message if encryption is enabled.
-    ///
-    /// - Parameter bytes: The bytes to encrypt.
-    /// - Returns: The encrypted bytes of encryption is enabled. Otherwise the input bytes.
-    private func encryptMessageIfRequired(_ bytes: ByteArray) -> ByteArray {
-        guard let messageCryptor = messageCryptor else {
-            return bytes
-        }
-
-        return Array(messageCryptor.encryptOutgoingMessage(Data(bytes: bytes)))
-    }
-
-    /// Decryptes a message if encryption is enabled.
-    ///
-    /// - Parameter bytes: The bytes to encrypt.
-    /// - Returns: The decrypted bytes of encryption is enabled. Otherwise the input bytes.
-    private func decryptMessageIfRequired(_ bytes: ByteArray) -> ByteArray {
-        guard let messageCryptor = messageCryptor else {
-            return bytes
-        }
-
-        return Array(messageCryptor.decryptIncommingMessage(Data(bytes: bytes)))
-    }
-}
-
-// MARK: - MinecraftClient+Compression
-extension MinecraftClient {
-    /// Compresses a message if compression is enabled.
-    ///
-    /// - Parameter bytes: The bytes to compress.
-    /// - Returns: The compressed bytes of compression is enabled. Otherwise the input bytes.
-    /// - Throws: And error which might occure.
-   private  func compressMessageIfRequired(_ bytes: ByteArray) throws -> ByteArray {
-        return try compressor?.compressMessage(bytes) ?? bytes
-    }
-
-    /// Decompresses a message if compression is enabled.
-    ///
-    /// - Parameter bytes: The bytes to decompress.
-    /// - Returns: The decompressed bytes of compression is enabled. Otherwise the input bytes.
-    /// - Throws: And error which might occure.
-    private func decompressMessageIfRequired(_ bytes: ByteArray) throws -> ByteArray {
-        return try compressor?.decompressMessage(bytes) ?? bytes
     }
 }
